Add Sangiin models and reuse in utils
