+++ src/models.py
@@ -55,3 +55,19 @@ class ShuShitsumonStatus(BaseModel):
     withdrawal_date: Optional[str]
     withdrawal_notice_date: Optional[str]
     status: Optional[str]
+
+
+class SangiinShitsumonData(BaseModel):
+    number: Optional[int] = Field(None)
+    question_subject: Optional[str] = Field(None)
+    submitter_name: Optional[str] = Field(None)
+    question_html_link: Optional[str] = Field(None)
+    question_pdf_link: Optional[str] = Field(None)
+    answer_html_link: Optional[str] = Field(None)
+    answer_pdf_link: Optional[str] = Field(None)
+
+
+class SangiinShitsumonList(BaseModel):
+    session: int
+    source: str
+    items: List[SangiinShitsumonData]
